import hugo 2.08 pre-alpha
https://www.zeograd.com/download/hugo208pas.tgz

sha256sum hugo208pas.tgz
472b9429d99d8000d54d652bb9ef581f91fc8e6efd5cdc2bcfedc6f3239d0033  hugo208pas.tgz

release notes:
This is the source of the first version of hugo with no allegro inside.
You'll need the runtime environnement available at http://www.libsdl.org .
I used version 1.2.2 but I guess that you can use an older version,
providing that you manually create a link so that the librairie loader
will be fooled.
This version doesn't have sound, nor all control keys, cd support is
untested, you have to set the filename of the rom on the command line
and you'll have to kill the emulator by force (like with control \)
but ... hey, it's an allegro free binary :)