pcecd.c: Add some helper macros
Making macros to substitute a few frequently used commands. They perform important functions, and are called multiple times within the code, but the actual function has not been described in the code.

I felt a macro was the ideal way to handle this, since adding comments to each instance of these commands would increase clutter.

This will make the code easier to read, and also easier to maintain by reducing code duplication.